+++ core/src/test/resources/flows/valids/working-directory.yaml
@@ -10,11 +10,11 @@ tasks:
     type: io.kestra.plugin.core.flow.WorkingDirectory
     tasks:
       - id: first
-        type: io.kestra.plugin.core.storage.LocalFiles
+        type: io.kestra.plugin.core.flow.WorkingDirectoryTest$LocalFiles
         inputs:
           out/stay.txt: "{{ inputs.failed == 'true' ? ko: taskrun.id }}"
       - id: second
-        type: io.kestra.plugin.core.storage.LocalFiles
+        type: io.kestra.plugin.core.flow.WorkingDirectoryTest$LocalFiles
         outputs:
           - out/*
       - id: disabled
